moved secret waypoints to its own file changed SecretPickupEvent stuff to be more like ClickEvent moved event dispatching to EventDispatcher
