+++ odinmain/src/main/kotlin/me/odinmain/OdinMain.kt
@@ -6,6 +6,7 @@ import me.odinmain.commands.registerCommands
 import me.odinmain.config.*
 import me.odinmain.events.EventDispatcher
 import me.odinmain.features.ModuleManager
+import me.odinmain.features.impl.dungeon.dungeonwaypoints.SecretWaypoints
 import me.odinmain.features.impl.render.*
 import me.odinmain.font.OdinFont
 import me.odinmain.ui.clickgui.ClickGUI
@@ -16,7 +17,6 @@ import me.odinmain.utils.render.*
 import me.odinmain.utils.sendDataToServer
 import me.odinmain.utils.skyblock.*
 import me.odinmain.utils.skyblock.dungeon.DungeonUtils
-import me.odinmain.utils.skyblock.dungeon.SecretUtils
 import net.minecraft.client.Minecraft
 import net.minecraft.client.gui.GuiScreen
 import net.minecraftforge.common.MinecraftForge
@@ -52,7 +52,6 @@ object OdinMain {
             WaypointManager,
             DevPlayers,
             SkyblockPlayer,
-            SecretUtils,
             //HighlightRenderer,
             //OdinUpdater,
             this
